First commit
Added .gitignore, constants.py and main2.py

main2.py is named so because I did a main.py as a test version (which
works fine) and I wish to keep it private, mainly because there isn’t a
single comment or docstring and it’s ugly as hell.

So here we are:

You can download any story from fan fiction.net (probably, I didn’t
really tried that far, just with a few of them), just enter the link of
the first chapter and it should works fine. If it don’t, report it
please. ;)