Version 3
The third version of the ffn_downloader is now fully completed.

As before, constants are separated from the rest in they own file.

I didn’t make a file for each class plus a ‘main’ file because I don’t
think they are long enough to do so. Not even 400 rows, counting the
full documentation in that, that didn’t seemed long enough to justify
splitting all of this and double the number of files.

REQUIREMENTS:
 • Python 3, if possible the latest version.
 • BeautifulSoup4 is now needed (package: bs4)

FUNCTIONALITIES:
    OLD:
 • Downloading a multiple chapters story,
 • Making one file per chapter,
 • Making one file containing the full story,
 • Making one file containing the informations about the story.
    NEW:
 + Downloading one-shot stories,
 + Each chapter now contains links to access the previous/next one and
those links aren’t included when not needed (first/last chapters and
one-shot stories),
 + The full file of the story now contains a table of contents to
easily access each chapter with internal anchors placed on the
chapters’s title.
 + A previously downloaded story can now be updated if a chapter was
added since the last without having to download the entire story again.
Obviously, the informations file and the full story are updated in the
process.
 + Some security checks were added to ensure the story is correctly
downloaded.
 + Several changes in the naming pattern should make it easier to find
the story you’re looking for if you downloaded lots of them (like me).

About the changes in the naming pattern:
 • The directory containing the story is now named (text-id)_(num-id),
to allow the OS to sort them alphabetically.
 • Each chapter is now named (zeros)(chap_num).html to ease the
creation of the links from chapter to chapter.
 • The full story file follow the same pattern as the directory:
(text-id)_(num-id).html.
 • The informations file follow the pattern (text-id)_infos.html.

And the last BIG piece of news:

I did with classes !

Yes, I know they are probably ugly and all, but I love them and it
really helped making the code as clear as possible at my level,
especially using objects.

I like to think I did fine, but really, I don’t know. It works so it
should be mostly fine though.