Infos update
REAL INFOS UPDATE (I messed up with the previous one)

Added a new variable in Story:

`universe` represents the universe in which the story is played (Harry
Potter for example). It includes crossover too.

Deleted Story.infos, because I never use it anywhere in my program (and
in my other files either).

To get the universe, the way to get the informations was slightly
changed.

Due to some bugs I encountered in the HTML of some stories, I added a
new constant and used it in correct_html to fix those bugs.

I modified the display a little to display the remaining number of
chapter to process.

Of course, I added the universe in the infos file, modifying it a
little.

I’m working on a stats.py file which should greatly help you compile
stats about all the stories you downloaded until now (or plan to
download, it depends how you’ll use it). It’s working since quite a
while but it’s so ugly and unorganized that I’m too ashamed to upload
it.