Corrections of bugs.
`stats.py`:
Fixed a missing spaces which caused the number of stories per
universe to be glued to 'stories'

`worker.py`:
Fixed a bug when the paths used in `OnComputer.update_stats` were
relative paths and not accessible from the current working directory.
A private method `get_full_path(self, path: str)` was added to
transform relative paths in full paths if needed.

None of this changes should affect the use of the program and code
working actually should still work after this update.

The example was updated to reflect this changes.