Worker Update
Some slight changes to `worker.py` that I feel were needed for it to be
easier to use (or less headache-inducing depending on your point of
view).

The main problem was that, when updating the statistics of a particular
folder (via the `update_stats` function), if `main` was `True` it was
supposed to also update the main stat-file. It dit, but not as intended
because the program didn’t know which folder(s) to use to build its
statistics, leading to errors. To avoid that, it is now simply disabled
by default and if you  want to update the main statistics, you will
have to be clear and explicitly provide all the folders to include.