Links update
+ Added: a AUTHOR_ID_REGEX to get the story’s author’s ID.
+ Added: a `Story.ifs['auth_id']` value which stores the author’s ID.

This ID is used to create links to the author’s page, in the
informations page and in each chapters, at the top and the bottom of
them.

Likewise, a link to the informations page was added using the story’s
title at the top and bottom of each chapter.

A section on "How to read a story" was added in the 'README.md' to
explain what was there before and what is there now too. I hope I
didn’t forget anything important.